Update upload.html
If we observe the user interface of the upload to Wikimedia commons page, it is not clear that the metadata fields can be edited. 
Hence a change in the html code of the upload to Wikimedia commons page is proposed to inform the user that they can edit the metadata fields. A text note saying "metadata fields are editable" is displayed in the left side just below the line "your selected file(s)".